Aİ/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
openai>=1.12.0
PyPDF2==3.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
yt-dlp>=2024.1.1
textblob==0.17.1

